govtender-scout/backend/requirements.txt
# GovTender Scout - Backend Dependencies (Free Tier Only)

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Scraping (free, no proxy required for basic usage)
playwright==1.44.0
beautifulsoup4==4.12.3
httpx==0.27.0

# PDF Processing (open source)
PyMuPDF==1.24.3
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0

# Task Scheduling
apscheduler==3.10.4
redis==5.0.4

# Auth (Supabase free tier)
supabase==2.4.6

# Payments (Razorpay SDK - free to integrate)
razorpay==1.4.1

# Utilities
python-dotenv==1.0.1
